+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/seasons/Seasons.java
@@ -1,6 +1,5 @@
 package insane96mcp.iguanatweaksreborn.module.world.seasons;
 
-import com.google.common.collect.Lists;
 import insane96mcp.iguanatweaksreborn.InsaneSO;
 import insane96mcp.iguanatweaksreborn.data.generator.ISOBlockTagsProvider;
 import insane96mcp.iguanatweaksreborn.event.HookTickToHookLureEvent;
@@ -13,16 +12,12 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Holder;
-import net.minecraft.core.SectionPos;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.placement.VegetationPlacements;
 import net.minecraft.server.level.ChunkHolder;
-import net.minecraft.server.level.ChunkMap;
-import net.minecraft.server.level.DistanceManager;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.TagKey;
-import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.GameRules;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LightLayer;
@@ -31,13 +26,13 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.DoublePlantBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.LevelChunk;
-import net.minecraft.world.level.chunk.LevelChunkSection;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.configurations.RandomPatchConfiguration;
 import net.minecraft.world.level.levelgen.placement.PlacedFeature;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.server.ServerStartedEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
 import sereneseasons.api.season.Season;
@@ -48,7 +43,6 @@ import sereneseasons.season.SeasonHandler;
 import sereneseasons.season.SeasonSavedData;
 import sereneseasons.season.SeasonTime;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -87,8 +81,8 @@ public class Seasons extends Feature {
 	@Config
 	public static Boolean seasonBasedFishingTime = true;
 
-	public Seasons(Module module, boolean enabledByDefault, boolean canBeDisabled) {
-		super(module, enabledByDefault, canBeDisabled);
+	public void init(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+		super.init(module, enabledByDefault, canBeDisabled);
 		InsaneSO.addServerPack("serene_seasons_changes", "Insane's Survival Overhaul Serene Seasons Changes", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && noGreenHouseGlass);
 		InsaneSO.addServerPack("no_saplings_in_winter", "Insane's Survival Overhaul No Saplings in Winter", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && noSaplingsInWinter);
 		if (ModList.get().isLoaded("tide"))
@@ -145,87 +139,6 @@ public class Seasons extends Feature {
 		}
 	}
 
-	@SubscribeEvent
-	public void onLevelTick(TickEvent.LevelTickEvent event) {
-        if (!this.isEnabled()
-				|| !grassDecayAndGrowth
-				|| event.phase != TickEvent.Phase.END
-				|| event.side != LogicalSide.SERVER
-				|| !event.level.getGameRules().getBoolean(RULE_SEASONGRASSGROWDEATH))
-            return;
-
-        Season.SubSeason subSeason = SeasonHelper.getSeasonState(event.level).getSubSeason();
-		Optional<GrassTickingData> data = GrassTickingData.get(subSeason);
-		if (data.isEmpty() || !data.get().shouldTick(event.level))
-			return;
-
-        ServerLevel level = (ServerLevel)event.level;
-		level.getProfiler().push("tallGrassRandomTick");
-        ChunkMap chunkMap = level.getChunkSource().chunkMap;
-        DistanceManager distanceManager = chunkMap.getDistanceManager();
-        int naturalSpawnChunkCount = distanceManager.getNaturalSpawnChunkCount();
-        List<ChunkAndHolder> list = Lists.newArrayListWithCapacity(naturalSpawnChunkCount);
-        chunkMap.getChunks().forEach(chunkHolder -> {
-            LevelChunk levelChunk = chunkHolder.getTickingChunk();
-            if (levelChunk != null)
-				list.add(new ChunkAndHolder(levelChunk, chunkHolder));
-        });
-
-        Collections.shuffle(list);
-		for (ChunkAndHolder chunkAndHolder : list) {
-			ChunkPos chunkPos = chunkAndHolder.chunk.getPos();
-			if (level.shouldTickBlocksAt(chunkPos.toLong()) && (chunkMap.anyPlayerCloseEnoughForSpawning(chunkPos) || distanceManager.shouldForceTicks(chunkPos.toLong()))) {
-				tickPlantsLifeDeath(chunkMap, chunkAndHolder.chunk, data.get());
-			}
-		}
-		level.getProfiler().pop();
-    }
-
-	private static void tickPlantsLifeDeath(ChunkMap chunkMap, LevelChunk levelChunk, GrassTickingData data) {
-		ServerLevel level = chunkMap.level;
-		ChunkPos chunkpos = levelChunk.getPos();
-		int x = chunkpos.getMinBlockX();
-		int z = chunkpos.getMinBlockZ();
-
-		LevelChunkSection[] levelChunkSections = levelChunk.getSections();
-		for (int s = 0; s < levelChunkSections.length; s++) {
-			LevelChunkSection levelchunksection = levelChunkSections[s];
-			if (levelchunksection.isRandomlyTicking()) {
-				int sectionY = levelChunk.getSectionYFromSectionIndex(s);
-				int y = SectionPos.sectionToBlockCoord(sectionY);
-
-				int randomTickSpeed = level.getGameRules().getInt(GameRules.RULE_RANDOMTICKING);
-				for (int t = 0; t < randomTickSpeed; t++) {
-					BlockPos pos = level.getBlockRandomPos(x, y, z, 15);
-					BlockState state = levelchunksection.getBlockState(pos.getX() - x, pos.getY() - y, pos.getZ() - z);
-					BlockPos abovePos = pos.above();
-					if (level.getBrightness(LightLayer.SKY, abovePos) < data.lightLevel)
-						continue;
-					if (state.is(BlockTags.DIRT)) {
-						BlockState stateUp = level.getBlockState(abovePos);
-						if (level.getRandom().nextInt(data.chance) == 0) {
-							if (stateUp.is(PLANTS_TO_DECAY))
-								level.setBlock(abovePos, Blocks.AIR.defaultBlockState(), 3);
-							else if (stateUp.is(PLANTS_TO_DEAD_BUSH))
-								level.setBlock(abovePos, Blocks.DEAD_BUSH.defaultBlockState(), 3);
-						}
-						else if (state.is(Blocks.GRASS_BLOCK) && level.getRandom().nextInt(data.chance) == 0 && stateUp.isAir()) {
-							Optional<Holder.Reference<PlacedFeature>> oPlacedFeature = level.registryAccess().registryOrThrow(Registries.PLACED_FEATURE).getHolder(VegetationPlacements.GRASS_BONEMEAL);
-							oPlacedFeature.ifPresent(placedFeatureReference ->
-									placedFeatureReference.value().place(level, level.getChunkSource().getGenerator(), level.random, abovePos));
-						}
-					}
-					else if (data.canGrowTall && level.getRandom().nextInt(data.chance) == 0 && level.getBlockState(pos.above()).canBeReplaced()) {
-						if (state.is(Blocks.GRASS))
-							DoublePlantBlock.placeAt(level, Blocks.TALL_GRASS.defaultBlockState(), pos, 2);
-						else if (state.is(Blocks.FERN))
-							DoublePlantBlock.placeAt(level, Blocks.LARGE_FERN.defaultBlockState(), pos, 2);
-					}
-				}
-			}
-		}
-	}
-
 	public static void tickPlantLifeDeath(BlockState state, BlockPos pos, ServerLevel level) {
 		Season.SubSeason subSeason = SeasonHelper.getSeasonState(level).getSubSeason();
 		Optional<GrassTickingData> oData = GrassTickingData.get(subSeason);
@@ -256,26 +169,36 @@ public class Seasons extends Feature {
 			else if (state.is(Blocks.FERN))
 				DoublePlantBlock.placeAt(level, Blocks.LARGE_FERN.defaultBlockState(), pos, 2);
 		}
+		else if (data.canGrowFlower
+				&& (state.is(Blocks.GRASS) || state.is(Blocks.FERN))
+				&& level.getRandom().nextInt(data.chance * 10) == 0) {
+			List<ConfiguredFeature<?, ?>> list = level.getBiome(pos).value().getGenerationSettings().getFlowerFeatures();
+			if (!list.isEmpty()) {
+				Holder<PlacedFeature> holder = ((RandomPatchConfiguration)list.get(0).config()).feature();
+				level.setBlock(pos, Blocks.AIR.defaultBlockState(), 3);
+				holder.value().place(level, level.getChunkSource().getGenerator(), level.random, pos);
+			}
+		}
 	}
 
 	record ChunkAndHolder(LevelChunk chunk, ChunkHolder holder) {}
 
 	public static final List<GrassTickingData> GRASS_TICKING_DATA = List.of(
-			new GrassTickingData(Season.SubSeason.EARLY_SPRING, 1100, GrassTickingType.GROW, 12, false),
-			new GrassTickingData(Season.SubSeason.MID_SPRING, 1100, GrassTickingType.GROW, 12, false),
-			new GrassTickingData(Season.SubSeason.LATE_SPRING, 550, GrassTickingType.GROW, 10, false),
-			new GrassTickingData(Season.SubSeason.EARLY_SUMMER, 275, GrassTickingType.GROW, 8, true),
-			new GrassTickingData(Season.SubSeason.MID_SUMMER, 275, GrassTickingType.GROW, 7, true),
-			new GrassTickingData(Season.SubSeason.LATE_SUMMER, 275, GrassTickingType.GROW, 7, true),
-			new GrassTickingData(Season.SubSeason.EARLY_AUTUMN, 2000, GrassTickingType.DECAY, 0, false),
-			new GrassTickingData(Season.SubSeason.MID_AUTUMN, 1000, GrassTickingType.DECAY, 0, false),
-			new GrassTickingData(Season.SubSeason.LATE_AUTUMN, 500, GrassTickingType.DECAY, 0, false),
-			new GrassTickingData(Season.SubSeason.EARLY_WINTER, 200, GrassTickingType.DECAY, 0, false),
-			new GrassTickingData(Season.SubSeason.MID_WINTER, 100, GrassTickingType.DECAY, 0, false),
-			new GrassTickingData(Season.SubSeason.LATE_WINTER, 400, GrassTickingType.DECAY, 0, false)
+			new GrassTickingData(Season.SubSeason.EARLY_SPRING, 1100, GrassTickingType.GROW, 12, false, false),
+			new GrassTickingData(Season.SubSeason.MID_SPRING, 1100, GrassTickingType.GROW, 12, false, true),
+			new GrassTickingData(Season.SubSeason.LATE_SPRING, 550, GrassTickingType.GROW, 10, false, true),
+			new GrassTickingData(Season.SubSeason.EARLY_SUMMER, 275, GrassTickingType.GROW, 8, true, true),
+			new GrassTickingData(Season.SubSeason.MID_SUMMER, 275, GrassTickingType.GROW, 7, true, true),
+			new GrassTickingData(Season.SubSeason.LATE_SUMMER, 275, GrassTickingType.GROW, 7, true, false),
+			new GrassTickingData(Season.SubSeason.EARLY_AUTUMN, 2000, GrassTickingType.DECAY, 0, false, false),
+			new GrassTickingData(Season.SubSeason.MID_AUTUMN, 1000, GrassTickingType.DECAY, 0, false, false),
+			new GrassTickingData(Season.SubSeason.LATE_AUTUMN, 500, GrassTickingType.DECAY, 0, false, false),
+			new GrassTickingData(Season.SubSeason.EARLY_WINTER, 200, GrassTickingType.DECAY, 0, false, false),
+			new GrassTickingData(Season.SubSeason.MID_WINTER, 100, GrassTickingType.DECAY, 0, false, false),
+			new GrassTickingData(Season.SubSeason.LATE_WINTER, 400, GrassTickingType.DECAY, 0, false, false)
 	);
 
-	public record GrassTickingData(Season.SubSeason subSeason, int chance, GrassTickingType grassTickingType, int lightLevel, boolean canGrowTall) {
+	public record GrassTickingData(Season.SubSeason subSeason, int chance, GrassTickingType grassTickingType, int lightLevel, boolean canGrowTall, boolean canGrowFlower) {
 		public boolean shouldTick(Level level) {
 			return chance > 0 && (grassTickingType == GrassTickingType.DECAY || level.isDay());
 		}